+++ src/utils.ts
@@ -17,7 +17,7 @@ export function normalizePath(id: string): string {
   return path.posix.normalize(os.platform() === 'win32' ? slash(id) : id)
 }
 
-export function normalizePathResolve(p1, p2) {
+export function normalizePathResolve(p1: string, p2: string) {
   return normalizePath(path.resolve(p1, p2))
 }
 
@@ -27,20 +27,17 @@ export const normalizeCssFilename = (p: string) =>
   p.replace(/\.(less|scss)/, '.css')
 
 export function relaceCssUrlPrefix(code: string) {
-  if (typeof code === 'string') {
-    return code.replace(/(?<=url\()[\s\S]*?(?=\))/gm, function (str) {
-      return (
-        'chrome-extension://' +
-        slash(path.join('__MSG_@@extension_id__', str.trim()))
-      )
-    })
-  }
-  return code
+  return code.replace(/(?<=url\()[\s\S]*?(?=\))/gm, function (str) {
+    return (
+      'chrome-extension://' +
+      slash(path.join('__MSG_@@extension_id__', str.trim()))
+    )
+  })
 }
 
 export function relaceResourcePathPrefix(code: string) {
   return code.replace(
-    /(?<=(=))"[^(?!")]+(\.png|jpg|jpeg|svg|webp|gif|mp3|mp4|avi|rmvb|mpeg|ra|ram|mov|wmv)"(?=,|;)/gm,
+    /(?<=(=))"[^(?!")]+(\.png|jpg|jpeg|svg|webp|gif|mp3|mp4|avi|rmvb|mpeg|ra|ram|mov|wmv|pdf)"(?=,|;)/gm,
     function (str) {
       return str.startsWith('http') ? str : `chrome.runtime.getURL(${str})`
     }
@@ -54,3 +51,5 @@ export function convertIntoIIFE(code: string) {
 export function isObject(value: unknown): value is Record<string, any> {
   return Object.prototype.toString.call(value) === '[object Object]'
 }
+
+export const isString = (value: unknown): value is string => typeof value === 'string'