+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.6 (2023-2-9)
++ fix:  Rollup input issue
++ chore: Overall code optimization
+
 ## 0.0.5 (2022-12-26)
 + fix: Page does not reload after modifying content_scripts's sass/less.
 + fix: Cannot remove dynamically injected CSS.